Struct crc::Crc

pub struct Crc<W: Width, I: Implementation = Table<1>> {
    pub algorithm: &'static Algorithm<W>,
    /* private fields */
}

Crc instance with a specific width, algorithm, and implementation.

Fields
algorithm: &'static Algorithm<W>
Implementations

impl Crc<u128, Table<1>>

pub const fn new(algorithm: &'static Algorithm<u128>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u128

pub const fn digest(&self) -> Digest<'_, u128, Table<1>>

pub const fn digest_with_initial( &self, initial: u128, ) -> Digest<'_, u128, Table<1>>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.

pub const fn table(&self) -> &<Table<1> as Implementation>::Data<u128>

impl Crc<u128, NoTable>

pub const fn new(algorithm: &'static Algorithm<u128>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u128

pub const fn digest(&self) -> Digest<'_, u128, NoTable>

pub const fn digest_with_initial( &self, initial: u128, ) -> Digest<'_, u128, NoTable>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.

impl Crc<u128, Table<16>>

pub const fn new(algorithm: &'static Algorithm<u128>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u128

pub const fn digest(&self) -> Digest<'_, u128, Table<16>>

pub const fn digest_with_initial( &self, initial: u128, ) -> Digest<'_, u128, Table<16>>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.

pub const fn table(&self) -> &<Table<16> as Implementation>::Data<u128>

impl Crc<u16, Table<1>>

pub const fn new(algorithm: &'static Algorithm<u16>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u16

pub const fn digest(&self) -> Digest<'_, u16, Table<1>>

pub const fn digest_with_initial( &self, initial: u16, ) -> Digest<'_, u16, Table<1>>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.

pub const fn table(&self) -> &<Table<1> as Implementation>::Data<u16>

impl Crc<u16, NoTable>

pub const fn new(algorithm: &'static Algorithm<u16>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u16

pub const fn digest(&self) -> Digest<'_, u16, NoTable>

pub const fn digest_with_initial( &self, initial: u16, ) -> Digest<'_, u16, NoTable>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.

impl Crc<u16, Table<16>>

pub const fn new(algorithm: &'static Algorithm<u16>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u16

pub const fn digest(&self) -> Digest<'_, u16, Table<16>>

pub const fn digest_with_initial( &self, initial: u16, ) -> Digest<'_, u16, Table<16>>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.

pub const fn table(&self) -> &<Table<16> as Implementation>::Data<u16>

impl Crc<u32, Table<1>>

pub const fn new(algorithm: &'static Algorithm<u32>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u32

pub const fn digest(&self) -> Digest<'_, u32, Table<1>>

pub const fn digest_with_initial( &self, initial: u32, ) -> Digest<'_, u32, Table<1>>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.

pub const fn table(&self) -> &<Table<1> as Implementation>::Data<u32>

impl Crc<u32, NoTable>

pub const fn new(algorithm: &'static Algorithm<u32>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u32

pub const fn digest(&self) -> Digest<'_, u32, NoTable>

pub const fn digest_with_initial( &self, initial: u32, ) -> Digest<'_, u32, NoTable>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.

impl Crc<u32, Table<16>>

pub const fn new(algorithm: &'static Algorithm<u32>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u32

pub const fn digest(&self) -> Digest<'_, u32, Table<16>>

pub const fn digest_with_initial( &self, initial: u32, ) -> Digest<'_, u32, Table<16>>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.

pub const fn table(&self) -> &<Table<16> as Implementation>::Data<u32>

impl Crc<u64, Table<1>>

pub const fn new(algorithm: &'static Algorithm<u64>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u64

pub const fn digest(&self) -> Digest<'_, u64, Table<1>>

pub const fn digest_with_initial( &self, initial: u64, ) -> Digest<'_, u64, Table<1>>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.

pub const fn table(&self) -> &<Table<1> as Implementation>::Data<u64>

impl Crc<u64, NoTable>

pub const fn new(algorithm: &'static Algorithm<u64>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u64

pub const fn digest(&self) -> Digest<'_, u64, NoTable>

pub const fn digest_with_initial( &self, initial: u64, ) -> Digest<'_, u64, NoTable>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.

impl Crc<u64, Table<16>>

pub const fn new(algorithm: &'static Algorithm<u64>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u64

pub const fn digest(&self) -> Digest<'_, u64, Table<16>>

pub const fn digest_with_initial( &self, initial: u64, ) -> Digest<'_, u64, Table<16>>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.

pub const fn table(&self) -> &<Table<16> as Implementation>::Data<u64>

impl Crc<u8, Table<1>>

pub const fn new(algorithm: &'static Algorithm<u8>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u8

pub const fn digest(&self) -> Digest<'_, u8, Table<1>>

pub const fn digest_with_initial(&self, initial: u8) -> Digest<'_, u8, Table<1>>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.

pub const fn table(&self) -> &<Table<1> as Implementation>::Data<u8>

impl Crc<u8, NoTable>

pub const fn new(algorithm: &'static Algorithm<u8>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u8

pub const fn digest(&self) -> Digest<'_, u8, NoTable>

pub const fn digest_with_initial(&self, initial: u8) -> Digest<'_, u8, NoTable>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.

impl Crc<u8, Table<16>>

pub const fn new(algorithm: &'static Algorithm<u8>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u8

pub const fn digest(&self) -> Digest<'_, u8, Table<16>>

pub const fn digest_with_initial( &self, initial: u8, ) -> Digest<'_, u8, Table<16>>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.

pub const fn table(&self) -> &<Table<16> as Implementation>::Data<u8>