Struct crc::Crc

pub struct Crc<W: Width> {
    pub algorithm: &'static Algorithm<W>,
    /* private fields */
}

Fields
algorithm: &'static Algorithm<W>
Implementations

impl Crc<u128>

pub const fn new(algorithm: &'static Algorithm<u128>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u128

pub const fn digest(&self) -> Digest<'_, u128>

pub const fn digest_with_initial(&self, initial: u128) -> Digest<'_, u128>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.

impl Crc<u16>

pub const fn new(algorithm: &'static Algorithm<u16>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u16

pub const fn digest(&self) -> Digest<'_, u16>

pub const fn digest_with_initial(&self, initial: u16) -> Digest<'_, u16>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.

impl Crc<u32>

pub const fn new(algorithm: &'static Algorithm<u32>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u32

pub const fn digest(&self) -> Digest<'_, u32>

pub const fn digest_with_initial(&self, initial: u32) -> Digest<'_, u32>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.

impl Crc<u64>

pub const fn new(algorithm: &'static Algorithm<u64>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u64

pub const fn digest(&self) -> Digest<'_, u64>

pub const fn digest_with_initial(&self, initial: u64) -> Digest<'_, u64>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.

impl Crc<u8>

pub const fn new(algorithm: &'static Algorithm<u8>) -> Self

pub const fn checksum(&self, bytes: &[u8]) -> u8

pub const fn digest(&self) -> Digest<'_, u8>

pub const fn digest_with_initial(&self, initial: u8) -> Digest<'_, u8>

Construct a Digest with a given initial value.

This overrides the initial value specified by the algorithm. The effects of the algorithm’s properties refin and width are applied to the custom initial value.